Revert "Apply the same tests to all storage backends (#453)"
This reverts commit c7c2cde0e6b7cb34f4fa307c1b8817af760dd184.
